requirements.txt
py-clob-client==0.34.6
python-dotenv==1.0.1
requests==2.32.3
websocket-client==1.8.0
web3>=7.0.0
structlog==24.4.0
pytest==8.3.2
pytest-asyncio==0.23.8
